requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
PyJWT==2.9.0